+++ jest.config.mjs
@@ -1,5 +1,6 @@
 // jest.config.mjs
 import nextJest from 'next/jest.js';
+import structuredClone from '@ungap/structured-clone';
 
 const createJestConfig = nextJest({
   // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
@@ -12,8 +13,11 @@ const config = {
   // Add more setup options before each test is run
   setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
 
-  // testEnvironment: 'jest-environment-jsdom',
-  testEnvironment: './FixJSDOMEnvironment.ts',
+  testEnvironment: 'jest-environment-jsdom',
+  // testEnvironment: './FixJSDOMEnvironment.ts',
+  globals: {
+    structuredClone: structuredClone,
+  },
 };
 
 // createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
