+++ package.json
@@ -32,6 +32,8 @@
     "@types/jest": "^29.5.1",
     "@types/react-pdf": "^6.2.0",
     "@types/testing-library__jest-dom": "^5.14.5",
+    "@types/ungap__structured-clone": "^0.3.0",
+    "@ungap/structured-clone": "^1.2.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0"
   }
